Delete and rename files
