+++ sdk/src/main/java/com/karhoo/sdk/api/service/quotes/QuotesService.kt
@@ -1,13 +1,10 @@
 package com.karhoo.sdk.api.service.quotes
 
 import com.karhoo.sdk.api.model.QuoteList
-import com.karhoo.sdk.api.model.QuoteListV2
 import com.karhoo.sdk.api.model.QuotesSearch
 import com.karhoo.sdk.call.PollCall
 
 interface QuotesService {
 
     fun quotes(quotesSearch: QuotesSearch): PollCall<QuoteList>
-
-    fun quotesV2(quotesSearch: QuotesSearch): PollCall<QuoteListV2>
 }
